move types to a separate file
